refactor: UserNickname 컴포넌트 분리 [TRL-54]
